Breaking out clientserver fixture module
